Add files via upload
Model to be used:
The data is formatted as standard CSV files. We will divide the data into training and test datasets and use Logistic regression/Decision tree algorithms to predict the whether an employee can be promoted or not.

Model effectiveness evaluation:
We will check the ROC curve, accuracy , confusion matrix and apply different models as per the data set- compare the accuracy of the algorithms applied.